Compile out NetworkAudioRecorder on Android
It requires the 'record audio' permission, which is always a bit scary to give.
Since there's no keybinds anyway to access it on Android, just disable so we can forego the permission altogether.

Note that currently EE *doesn't* request that permission: on most devices, this is silently ignored and the recording device just fails to open, but on some other it leads to a start up crash, which is what @StarryWisdom encountered.