Update scenario_44_outpost.lua (#1949)
Version 0.0.1 to 1.0.9
These are several modifications made by Xansta during the last 2 months after several tests and feedback that I made. Xansta just passed me the file so that it is up to date on the repository. So that I will be able standardize and complete the translation tags.
Modifications :
<<Described ranks in scenario description.
Corrected several typos in the ship upgrade descriptions (thanks muerte for finding them).
Added entry in science database explaining weapons banner with S, M and L for tube sizes.
Added some translation tags.>>